+++ lib/src/scope.dart
@@ -915,6 +915,9 @@ class Scope {
 
     _isDisposed = true;
 
+    // Call the blue print's onDispose method
+    bluePrint.onDispose(scope: this);
+
     // Dispose the scope's nodes
     for (final node in [..._nodes.values]) {
       node.dispose();